+++ src/server.ts
@@ -2,10 +2,10 @@ import * as http from 'http';
 import * as stream from 'stream';
 import {
   BadRequest,
+  Logger as BlessLogger,
   Config,
   HTTPRoute,
   Hooks,
-  Logger,
   Metrics,
   NotFound,
   Request,
@@ -50,6 +50,7 @@ export class HTTPServer extends EventEmitter {
     protected token: Token,
     protected router: Router,
     protected hooks: Hooks,
+    protected Logger: typeof BlessLogger,
   ) {
     super();
     this.host = config.getHost();
@@ -264,7 +265,7 @@ export class HTTPServer extends EventEmitter {
     }
 
     return (route as HTTPRoute)
-      .handler(req, res, new Logger(route.name, req))
+      .handler(req, res, new this.Logger(route.name, req))
       .then(() => {
         this.verbose('HTTP connection complete');
       })
@@ -369,7 +370,7 @@ export class HTTPServer extends EventEmitter {
       }
 
       return (route as WebSocketRoute)
-        .handler(req, socket, head, new Logger(route.name, req))
+        .handler(req, socket, head, new this.Logger(route.name, req))
         .then(() => {
           this.verbose('Websocket connection complete');
         })